// src/enums/rangedWeaponName.ts
export enum RangedWeaponName {
	Fidelis = 'Fidelis',
	ParagonMissileLauncherPrioris = 'Paragon missile launcher – prioris',
	ParagonMissileLauncherSanctorum = 'Paragon missile launcher – sanctorum',
	BoltPistol = 'Bolt pistol',
	BrazierOfHolyFire = 'Brazier of holy fire',
	CondemnorBoltgun = 'Condemnor boltgun',
	InfernoPistol = 'Inferno pistol',
	PlasmaPistolStandard = 'Plasma pistol – standard',
	PlasmaPistolSupercharge = 'Plasma pistol – supercharge',
	TwinMinistorumHeavyFlamer = 'Twin Ministorum heavy flamer',
	Autogun = 'Autogun',
	HolyPistol = 'Holy pistol',
	MinistorumShotgun = 'Ministorum shotgun',
	TheArdentBlade = 'The Ardent Blade',
	BoltPistols = 'Bolt pistols',
	ArtificerCraftedStormBolter = 'Artificer-crafted storm bolter',
	Boltgun = 'Boltgun',
	HeavyBolter = 'Heavy bolter',
	Meltagun = 'Meltagun',
	MinistorumFlamer = 'Ministorum flamer',
	MinistorumHeavyFlamer = 'Ministorum heavy flamer',
	MultiMelta = 'Multi-melta',
	Autopistol = 'Autopistol',
	ZealotsVindictor = 'Zealot’s vindictor',
	MinistorumHandFlamer = 'Ministorum hand flamer',
	ParagonGrenadeLaunchers = 'Paragon grenade launchers',
	ParagonStormBolters = 'Paragon storm bolters',
	MortifierFlamer = 'Mortifier flamer',
	PenitentFlamers = 'Penitent flamers',
	ExorcistConflagrationRockets = 'Exorcist conflagration rockets',
	ExorcistMissileLauncher = 'Exorcist missile launcher',
	HunterKillerMissile = 'Hunter-killer missile',
	CastigatorAutocannons = 'Castigator autocannons',
	CastigatorBattleCannon = 'Castigator battle cannon',
	StormBolter = 'Storm bolter',
	ImmolationFlamers = 'Immolation flamers',
	TwinHeavyBolter = 'Twin heavy bolter',
	TwinMultiMelta = 'Twin multi-melta'
}